add grid sample phi kernel and remove original kernel
